text_extract/requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pdfminer.six>=20201018
pdf2image>=1.16.0
pytesseract>=0.3.8
python-docx>=0.8.11